Final changes for this refactor
