+++ Khana POC/src/utils/getWeb3.js
@@ -11,13 +11,20 @@ let getWeb3 = new Promise(function(resolve, reject) {
       // Use Mist/MetaMask's provider.
       web3 = new Web3(web3.currentProvider)
 
-      results = {
-        web3: web3
-      }
-      resolve(results)
+      web3.version.getNetwork((err, netId) => {
+        if(Number(netId) !== 4 && Number(netId) < 99) { // Let us work in development chains (usually > 100)
+          resolve(Error("Wrong network selected. We are only live on Rinkeby! Please switch 🙏"))
+        } else {
+          results = {
+            web3: web3
+          }
+          resolve(results)
+        }
+      })
+
     } else {
       // No web3 provider available!
-      resolve(Error("No Web3 provider detected! Try loading with a different browser or Metamask"))
+      resolve(Error("No Web3 provider detected! 😩 Try loading with a different browser or Metamask"))
     }
   })
 })
